OrderPage added, fixed errors, app finished.

+++ src/app/api/categories/route.js
@@ -1,17 +1,24 @@
 import { Category } from '@/models/Category'
 import mongoose from 'mongoose'
+import { isAdmin } from '../auth/[...nextauth]/route'
 
 export async function POST(req) {
   mongoose.connect(process.env.MONGO_URL)
-  const { name } = await req.json()
-  const categoryDoc = await Category.create({ name })
-  return Response.json(categoryDoc)
+  if (await isAdmin()) {
+    const { name } = await req.json()
+    const categoryDoc = await Category.create({ name })
+    return Response.json(categoryDoc)
+  } else {
+    return Response.json({})
+  }
 }
 
 export async function PUT(req) {
   mongoose.connect(process.env.MONGO_URL)
   const { _id, name } = await req.json()
-  await Category.updateOne({ _id }, { name })
+  if (await isAdmin()) {
+    await Category.updateOne({ _id }, { name })
+  }
   return Response.json(true)
 }
 
@@ -24,6 +31,8 @@ export async function DELETE(req) {
   mongoose.connect(process.env.MONGO_URL)
   const url = new URL(req.url)
   const _id = url.searchParams.get('_id')
-  await Category.deleteOne({ _id })
+  if (await isAdmin()) {
+    await Category.deleteOne({ _id })
+  }
   return Response.json(true)
 }
